# use bottleneck package to speedup some critical numpy operations.