--- a/requirements.txt
+++ b/requirements.txt
@@ -1,4 +1,5 @@
 numpy>=1.18
+bottleneck>=1.3.2
 pyepics>=3.4.0
 python-dateutil>=2.8.1
 requests>=2.22.0