checkout these performance tests for the cases I use in the code:
```
In [1]: import numpy as np

In [2]: import bottleneck as bn

In [3]: a = np.random.rand(20)

In [4]: %timeit np.nanmax(a)
5.17 µs ± 16 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [5]: %timeit bn.nanmax(a)
128 ns ± 0.26 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [6]: %timeit np.nanmin(a)
5.22 µs ± 150 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [7]: %timeit bn.nanmin(a)
126 ns ± 0.424 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [8]: b = np.random.rand(320)

In [9]: %timeit b.mean()
3.04 µs ± 24.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [10]: %timeit bn.nanmean(b)
399 ns ± 1.51 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [11]: %timeit b.std()
13 µs ± 17.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [12]: %timeit bn.nanstd(b)
701 ns ± 2.02 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [13]: %timeit b.max()
1.64 µs ± 8.01 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [14]: %timeit bn.nanmax(b)
377 ns ± 2.05 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [15]: %timeit b.min()
1.68 µs ± 7.03 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [16]: %timeit bn.nanmin(b)
290 ns ± 6.24 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

this is the bottleneck main page: https://pypi.org/project/Bottleneck/